Diagrammes uml24

+++ src/Controller/UserController.php
@@ -5,30 +5,86 @@ namespace App\Controller;
 use App\Entity\User;
 use App\Repository\ClientRepository;
 use App\Repository\UserRepository;
+use App\Service\UserService;
 use Doctrine\ORM\EntityManagerInterface;
+use JMS\Serializer\SerializationContext;
+use JMS\Serializer\SerializerInterface;
+use Nelmio\ApiDocBundle\Annotation\Model;
+use Psr\Cache\InvalidArgumentException;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
-use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
-use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
+use Symfony\Contracts\Cache\ItemInterface;
+use Symfony\Contracts\Cache\TagAwareCacheInterface;
+use OpenApi\Attributes as OA;
+
+
 
 class UserController extends AbstractController
 {
+
+    /**
+     * @param UserService $userService
+     */
+    public function __construct(private readonly UserService $userService)
+    {
+    }
+
     /**
      * @param UserRepository $userRepository
      * @param SerializerInterface $serializer
+     * @param Request $request
+     * @param TagAwareCacheInterface $cachePool
      * @return JsonResponse
+     * @throws InvalidArgumentException
      */
     #[Route('/api/users', name: 'users', methods: ['GET'])]
-    public function getAllUsers(UserRepository $userRepository, SerializerInterface $serializer): JsonResponse
+    #[OA\Response(
+        response: 200,
+        description: 'Retourne la liste des livres',
+        content: new OA\JsonContent(
+            type: 'array',
+            items: new OA\Items(ref: new Model(type: User::class, groups: ['getUsers']))
+        )
+    )]
+    #[OA\Parameter(
+        name: 'page',
+        description: "La page que l'on veut récupérer",
+        in: 'query',
+        schema: new OA\Schema(type: 'int')
+    )]
+    #[OA\Parameter(
+        name: 'limit',
+        description: "Le nombre d'éléments que l'on veut récupérer",
+        in: 'query',
+        schema: new OA\Schema(type: 'int')
+    )]
+    #[OA\Tag(name: 'Users')]
+    public function getAllUsers(UserRepository         $userRepository,
+                                SerializerInterface    $serializer,
+                                Request                $request,
+                                TagAwareCacheInterface $cachePool): JsonResponse
     {
-        $userList = $userRepository->findAll();
-        $jsonUserList = $serializer->serialize($userList, 'json', ['groups' => 'getUsers']);
+        $loggedInClient = $this->userService->validateLoggedInClient();
+
+        $page = $request->get('page', 1);
+        $limit = $request->get('limit', 3);
+
+        $idCache = "getAllUsers" . $page . "-" . $limit;
+
+        $userList = $cachePool->get($idCache,
+            function (ItemInterface $item) use ($userRepository, $page, $limit, $loggedInClient) {
+                $item->tag('UsersCache');
+                $item->expiresAfter(300);
+                return $userRepository->findUsersByClient($loggedInClient, $page, $limit);
+            });
+        $context = SerializationContext::create()->setGroups(['getUsers']);
+        $jsonUserList = $serializer->serialize($userList, 'json', $context);
 
         return new JsonResponse($jsonUserList, Response::HTTP_OK, [], true);
     }
@@ -39,10 +95,16 @@ class UserController extends AbstractController
      * @return JsonResponse
      */
     #[Route('/api/users/{id}', name: 'detailUser', methods: ['GET'])]
-    public function getDetailUser(User $user, SerializerInterface $serializer, Security $security): JsonResponse
+    public function getDetailUser(User $user, SerializerInterface $serializer): JsonResponse
     {
-        $client = $security->getUser();
-        $jsonUser = $serializer->serialize($user, 'json', ['groups' => 'getUsers']);
+        $validationResult = $this->userService->validateClientAccess($user);
+
+        if ($validationResult instanceof JsonResponse) {
+            return $validationResult;
+        }
+
+        $context = SerializationContext::create()->setGroups(['getUsers']);
+        $jsonUser = $serializer->serialize($user, 'json', $context);
 
         return new JsonResponse($jsonUser, Response::HTTP_OK, [], true);
     }
@@ -53,8 +115,14 @@ class UserController extends AbstractController
      * @return JsonResponse
      */
     #[Route('/api/users/{id}', name: 'deleteUser', methods: ['DELETE'])]
-    public function deleteMobile(User $user, EntityManagerInterface $em): JsonResponse
+    public function deleteUser(User $user, EntityManagerInterface $em): JsonResponse
     {
+        $validationResult = $this->userService->validateClientAccess($user);
+
+        if ($validationResult instanceof JsonResponse) {
+            return $validationResult;
+        }
+
         $em->remove($user);
         $em->flush();
 
@@ -67,64 +135,44 @@ class UserController extends AbstractController
      * @param EntityManagerInterface $em
      * @param ClientRepository $clientRepository
      * @param UrlGeneratorInterface $urlGenerator
+     * @param ValidatorInterface $validator
+     * @param UserPasswordHasherInterface $passwordHasher
      * @return JsonResponse
      */
     #[Route('/api/users', name: 'createUser', methods: ['POST'])]
-    public function createUser(Request                $request, SerializerInterface $serializer,
-                               EntityManagerInterface $em,
-                               ClientRepository       $clientRepository,
-                               UrlGeneratorInterface  $urlGenerator,
-                               ValidatorInterface     $validator): JsonResponse
+    public function createUser(Request                     $request,
+                               SerializerInterface         $serializer,
+                               EntityManagerInterface      $em,
+                               ClientRepository            $clientRepository,
+                               UrlGeneratorInterface       $urlGenerator,
+                               ValidatorInterface          $validator,
+                               UserPasswordHasherInterface $passwordHasher): JsonResponse
     {
+
         $user = $serializer->deserialize($request->getContent(), User::class, 'json');
 
         $errors = $validator->validate($user);
         if ($errors->count() > 0) {
             return new JsonResponse($serializer->serialize($errors, 'json'),
-                JsonResponse::HTTP_BAD_REQUEST, [], true);
+                Response::HTTP_BAD_REQUEST, [], true);
         }
 
         $content = $request->toArray();
         $idClient = $content["idClient"] ?? -1;
         $user->setClient($clientRepository->find($idClient));
+        $user->setRoles(['ROLE_USER']);
+        $plainPassword = $content['password'];
+        $user->setPassword($passwordHasher->hashPassword($user, $plainPassword));
 
         $em->persist($user);
         $em->flush();
 
-        $jsonUser = $serializer->serialize($user, 'json', ['groups' => 'getUsers']);
+        $context = SerializationContext::create()->setGroups(['getUsers']);
+        $jsonUser = $serializer->serialize($user, 'json', $context);
         $location = $urlGenerator->generate('detailUser', ['id' => $user->getId()],
             UrlGeneratorInterface::ABSOLUTE_URL);
 
         return new JsonResponse($jsonUser, Response::HTTP_CREATED, ['location' => $location], true);
-    }
-
-    /**
-     * @param Request $request
-     * @param SerializerInterface $serializer
-     * @param EntityManagerInterface $em
-     * @param ClientRepository $clientRepository
-     * @param User $currentUser
-     * @param UrlGeneratorInterface $urlGenerator
-     * @return JsonResponse
-     */
-    #[Route('/api/users/{id}', name: 'updateUser', methods: ['PUT'])]
-    public function updateUser(Request                $request, SerializerInterface $serializer,
-                               EntityManagerInterface $em,
-                               ClientRepository       $clientRepository,
-                               User                   $currentUser,
-                               UrlGeneratorInterface  $urlGenerator): JsonResponse
-    {
-        $updatedUser = $serializer->deserialize($request->getContent(), User::class, 'json',
-            [AbstractNormalizer::OBJECT_TO_POPULATE => $currentUser]);
-
-        $content = $request->toArray();
-        $idClient = $content["idClient"] ?? -1;
-        $updatedUser->setClient($clientRepository->find($idClient));
-
-        $em->persist($updatedUser);
-        $em->flush();
-
-        return new JsonResponse(null, JsonResponse::HTTP_NO_CONTENT);
     }
 
 }
